+++ src/main/java/com/nhnacademy/memberapi/dto/response/GradeResponse.java
@@ -2,11 +2,12 @@ package com.nhnacademy.memberapi.dto.response;
 
 import com.nhnacademy.memberapi.entity.Grade;
 import com.nhnacademy.memberapi.entity.GradeName;
+import java.math.BigDecimal;
 
 public record GradeResponse(
     Long gradeId,
     GradeName gradeName,
-    Integer gradePointRatio,
+    BigDecimal gradePointRatio,
     Integer gradeCondition
 ) {
     public static GradeResponse from(Grade grade) {
